Refactor: Move ProductDetailsDialogFragment to inventory package and update references
- Moved `ProductDetailsDialogFragment` to the `inventory` package.
- Updated package references to `ProductDetailsDialogFragment` to reflect the new location.
- Removed unused import from `InventoryAdapter`.
- Added view binding references `expiryInput` and `calendarIcon` to `AddManualProductDialogFragment`.
